fixing flask structure issues
