+++ frontstage/views/register.py
@@ -1,33 +1,22 @@
 import json
 import logging
-import os
 
-import requests
-from flask import Flask, Blueprint, render_template, request, flash, redirect, url_for, abort
+from flask import Blueprint, render_template, request, flash, redirect, url_for, abort
 from jose import JWTError
 from oauthlib.oauth2 import BackendApplicationClient, MissingTokenError
 from ons_ras_common import ons_env
+import requests
 from requests_oauthlib import OAuth2Session
 from structlog import wrap_logger
 
-from app.jwt import encode
-from app.models import RegistrationForm, EnrolmentCodeForm
 from config import Config
-from ..common.post_event import post_event
+from frontstage.common.post_event import post_event
+from frontstage.jwt import encode
+from frontstage.models import RegistrationForm, EnrolmentCodeForm
 
-app = Flask(__name__)
-app.debug = True
-
-app.config.update(
-    DEBUG=True,
-    TESTING=True,
-    TEMPLATES_AUTO_RELOAD=True
-)
-
-if 'APP_SETTINGS' in os.environ:
-    app.config.from_object(Config)
 
 logger = wrap_logger(logging.getLogger(__name__))
+
 register_bp = Blueprint('register_bp', __name__, static_folder='static', template_folder='templates/register')
 
 
@@ -108,14 +97,14 @@ def register_confirm_organisation_survey():
         decrypted_enrolment_code = ons_env.crypt.decrypt(encrypted_enrolment_code.encode()).decode()
     else:
         logger.error('Confirm organisation screen - Enrolment code not specified')
-        return redirect(url_for('error_page'))
+        return redirect(url_for('error_bp.error_page'))
 
     case_id = validate_enrolment_code(decrypted_enrolment_code)
 
     # Ensure we have got a valid enrolment code, otherwise go to the sign in page
     if not case_id:
         logger.error('Confirm organisation screen - Case ID not available')
-        return redirect(url_for('error_page'))
+        return redirect(url_for('error_bp.error_page'))
 
     # TODO More error handling e.g. cater for case not coming back from the case service, etc.
 
@@ -194,7 +183,7 @@ def register_enter_your_details():
     # Ensure we have got a valid enrolment code, otherwise go to the sign in page
     if not decrypted_enrolment_code or not validate_enrolment_code(decrypted_enrolment_code):
         logger.error('Enter Account Details page - invalid enrolment code: ' + str(decrypted_enrolment_code))
-        return redirect(url_for('error_page'))
+        return redirect(url_for('error_bp.error_page'))
 
     form = RegistrationForm(request.values, enrolment_code=encrypted_enrolment_code)
 
@@ -307,8 +296,8 @@ def register_enter_your_details():
         try:
             token = oauth.fetch_token(token_url=token_url, client_id=app.config['RAS_FRONTSTAGE_CLIENT_ID'],
                                       client_secret=app.config['RAS_FRONTSTAGE_CLIENT_SECRET'])
-            app.logger.debug(" *** Access Token Granted *** ")
-            app.logger.debug(" Values are: ")
+            logger.debug(" *** Access Token Granted *** ")
+            logger.debug(" Values are: ")
             for key in token:
                 logger.debug("{} Value is: {}".format(key, token[key]))
 
@@ -333,8 +322,8 @@ def register_enter_your_details():
             return abort(500, '{"message":"There was a problem with the Authentication service please contact a member of the ONS staff"}')
 
         except MissingTokenError as e:
-            app.logger.warning("Missing token error, error is: {}".format(e))
-            app.logger.warning("Failed validation")
+            logger.warning("Missing token error, error is: {}".format(e))
+            logger.warning("Failed validation")
 
             return abort(500, '{"message":"There was a problem with the Authentication service please contact a member of the ONS staff"}')
 
@@ -354,7 +343,7 @@ def register_enter_your_details():
 
         #party_service_url = app.config['API_GATEWAY_PARTY_URL'] + 'respondents'
         party_service_url = Config.RAS_PARTY_POST_RESPONDENTS.format(Config.RAS_PARTY_SERVICE)
-        app.logger.debug("Party service URL is: {}".format(party_service_url))
+        logger.debug("Party service URL is: {}".format(party_service_url))
 
         try:
             result = requests.post(party_service_url, headers=headers, data=json.dumps(registration_data))
@@ -364,11 +353,11 @@ def register_enter_your_details():
                 return render_template('register/register.almost-done.html', _theme='default', email=email_address)
             else:
                 logger.error('Unable to register user - Party service error user')
-                return redirect(url_for('error_page'))
+                return redirect(url_for('error_bp.error_page'))
 
         except ConnectionError:
             logger.critical("We could not connect to the party service")
-            return redirect(url_for('error_page'))
+            return redirect(url_for('error_bp.error_page'))
 
         # TODO We need to add an exception timeout catch and handle this type of error
 
@@ -409,12 +398,12 @@ def register_activate_account(token):
                 return redirect(url_for('register_bp.register_resend_email', user_id=user_id))
             else:
                 logger.error('Unable to determine user for activation token: ' + str(token))
-                return redirect(url_for('error_page'))
+                return redirect(url_for('error_bp.error_page'))
     else:
         # If the token was not recognised, we don't know who the user is so redirect them off to the error page
         logger.warning('Unrecognised email activation token: ' + str(token) +
                        ' Response code: ' + str(result.status_code))
-        return redirect(url_for('error_page'))
+        return redirect(url_for('error_bp.error_page'))
 
 
 @register_bp.route('/create-account/resend-email', methods=['GET'])
