+++ tests/test_registration.py
@@ -4,8 +4,9 @@ import unittest
 import requests_mock
 from ons_ras_common import ons_env
 
-from app.application import app
 from config import Config
+from frontstage import app
+
 
 with open('tests/test_data/my_surveys.json') as json_data:
     my_surveys_data = json.load(json_data)
